+++ src/Components/Attendees.js
@@ -1,3 +1,12 @@
-export default function Attendees() {
-  return;
+import Attendee from "./Attendee";
+
+export default function Attendees({attendee,event,updateEventAttendance}) {
+  return (
+    <Attendee 
+    attendee={attendee} 
+    event={event} 
+    updateEventAttendance={updateEventAttendance}
+    />
+  );
 }
+
