+++ src/Components/Header.js
@@ -1,3 +1,12 @@
 export default function Header() {
-  return;
+  return (
+    <>
+    <header>
+      <h1 className="color-change-5x">RSVP App</h1>
+    </header>
+  </>
+  );
 }
+
+
+
